create 2 folders and rearanged all my screens view
