+++ frontend/app/navigations/AuthNavigator.tsx
@@ -1,31 +1,25 @@
 import { createStackNavigator } from "@react-navigation/stack";
-import HomeScreen from "../screens/HomeScreen";
-import Signup from "../screens/signup";
+import HomeScreen from "../screens/unAuthScreens/HomeScreen";
+import Signup from "../screens/unAuthScreens/signup";
 import { View } from "react-native";
-import Login from "../screens/login";
-import ForgetPassword from "../screens/forgetPassword";
-import ReceiveSms from "../screens/receiveSms";
-import OtpVerification from "../screens/otpVerification";
+import Login from "../screens/unAuthScreens/login";
+import ForgetPassword from "../screens/unAuthScreens/forgetPassword";
+import ReceiveSms from "../screens/unAuthScreens/receiveSms";
+import OtpVerification from "../screens/unAuthScreens/otpVerification";
 import { createBottomTabNavigator } from "@react-navigation/bottom-tabs";
 import Icon from "react-native-vector-icons/Ionicons";
-import HomePage from "../screens/homePage";
-import Transaction from "../screens/transaction";
-import CreatePage from "../screens/createPage";
-import Stats from "../screens/stats";
-import Setting from "../screens/setting";
+import HomePage from "../screens/authScreens/homePage";
+import Transaction from "../screens/authScreens/transaction";
+import CreatePage from "../screens/authScreens/createPage";
+import Stats from "../screens/authScreens/stats";
+import Setting from "../screens/authScreens/setting";
 import TabNavigator, { RootTabParamList } from "./Tabnavigator";
 import { NavigatorScreenParams } from "@react-navigation/native";
 
 export type RootStackParamList = {
   ForgetPassword: undefined;
   "Ouvrir sms": {
-    name: string;
-    firstname: string;
     phone: string;
-    birthday: string;
-    statut: string;
-    password: string;
-    verificationId: any;
   };
   "Mot de passe oublie": undefined;
   login: undefined;
