+++ frontend/app/screens/authScreens/createPage.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { StyleSheet, View, Text } from "react-native";
 import { createStackNavigator } from "@react-navigation/stack";
-import Spend from "@/app/screens/createScreen/spend";
+import Spend from "@/app/screens/authScreens/createScreen/spend";
 import Tabs from "./createScreen/tabs";
 
 const CreatePage = () => {
