+++ frontend/app/screens/authScreens/createScreen/income.tsx
@@ -1,45 +1,45 @@
 import React from "react";
 import { StyleSheet, View, Text } from "react-native";
-import Button from "../../../components/Button";
+import Button from "../../../../components/Button";
 import * as zod from "zod";
 import { useForm, Controller } from "react-hook-form";
 import { zodResolver } from "@hookform/resolvers/zod";
-import InputComponent from "../../../components/inputComponent";
+import InputComponent from "../../../../components/inputComponent";
 
-const spendSchema = zod.object({
-  description: zod.string().min(1, "entrer la raison de la depense"),
+const incomeSchema = zod.object({
+  source: zod.string().min(1, "entrer la source de vos revenues svp"),
   montant: zod.coerce.number(),
 });
-type spendData = zod.infer<typeof spendSchema>;
-const Spend = () => {
+type incomeData = zod.infer<typeof incomeSchema>;
+const Income = () => {
   const {
     handleSubmit,
     control,
     formState: { errors },
-  } = useForm<spendData>({
-    resolver: zodResolver(spendSchema),
+  } = useForm<incomeData>({
+    resolver: zodResolver(incomeSchema),
   });
 
-  const onSubmit = (data: spendData) => {
+  const onSubmit = (data: incomeData) => {
     console.log("data", data);
   };
   return (
     <View className="containers w-full  px-4">
       <Text className="text-center font-helvitica-bold text-2xl w-full">
-        Creer une nouvelle depense
+        Creer une nouvelle source de revenue
       </Text>
       <View className="inputs w-full gap-4">
         <View className="desc gap-3 w-full">
           <Text className=" text-russian-950 text-lg font-raleway-medium">
-            Description
+            Source
           </Text>
           <Controller
             control={control}
-            name="description"
+            name="source"
             render={({ field: { onChange, onBlur, value } }) => (
               <InputComponent
                 type="default"
-                placeholder="Entrer un description"
+                placeholder="Entrer un source"
                 onChangeText={onChange}
                 isIcon={false}
                 className=""
@@ -71,4 +71,4 @@ const Spend = () => {
 
 const styles = StyleSheet.create({});
 
-export default Spend;
+export default Income;
